+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+- Global lock release on Ctrl+C during the operation
+
 ## [3.5.0]
 ### Added
 - Global locks support for content uploads, to prevent parallel uploads